pyproject.toml
[project]
name = "multims2"
version = "0.0.1"
description = "MultiMS2"
requires-python = ">=3.13,<4"
dependencies = [
    "altair-upset>=0.4.0,<1",
    "cmcrameri>=1.9,<2",
    "datasketch>=1.6.5,<2",
    "faerun>=0.4.7,<1",
    "lightgbm>=4.6.0,<5", # required by msbuddy
    "map4>=1.1.3,<2",
    "marimo>=0.16.5,<1",
    "matchms>=0.30.2,<1",
    "matplotlib-colors>=1.0.16",
    "molzip>=0.9.2,<1",
    "msbuddy @ git+https://github.com/adafede/msbuddy",
    "numpy>=2.2.6,<3",
    "polars>=1.34.0,<2",
    "pooch>=1.8.2,<2",
    "pre-commit>=4.3.0,<5",
    "pubchempy>=1.0.5,<2",
    "pyarrow>=21.0.0,<22",
    "rdkit>=2024.9.6,<2025",
    "scipy>=1.16.2,<2",
    "selfies>=2.2.0,<3",
    "simple-parsing>=0.1.7,<1",
    "tmap-viz @ git+https://github.com/mvisani/tmap",
    "tol-colors>=2.1.0,<3",
    "tqdm>=4.67.1,<5",
    "vl-convert-python>=1.8.0,<2",
]

[tool.deptry.per_rule_ignores]
DEP002 = ["pre-commit",]

[tool.ty.rules]
unresolved-import = "ignore"
